small refactoring in tests
