+++ tests/test_SOAPify.py
@@ -4,11 +4,10 @@ from numpy.random import randint
 from numpy.testing import assert_array_equal
 from SOAPify import SOAPReferences
 from ase.data import atomic_numbers
-from SOAPify.classify import SOAPclassification
 import pytest
 
 
-def test_atomicnumberOredring():
+def test_atomicnumberOrdering():
     species = ["O", "H"]
     assert SOAPify.orderByZ(species) == ["H", "O"]
     species = ["Au", "H", "C", "W", "O"]
@@ -61,38 +60,6 @@ def test_normalizeMatrix():
             sum = numpy.sum(newMatNorm[row, :])
             if sum != 0.0:
                 numpy.testing.assert_approx_equal(sum, 1.0, 8)
-
-
-@pytest.fixture(
-    scope="module",
-    params=[-1, 0, 1],
-)
-def shiftFixture(request):
-    return request.param
-
-
-nshift = shiftFixture
-lshift = shiftFixture
-
-
-def test_concatenationSOAPreferencesCompatibility(
-    nshift, lshift, lMaxFixture, nMaxFixture
-):
-    # lmax fixture can be 0, whit lshift -1 goes under 0, that is not  "matematically sensed",
-    # but this only tests if nmax and lmax are equal in the merge references, and not if the vectors are senses
-    a = SOAPReferences(["a", "b"], [[0, 0], [1, 1]], nMaxFixture, lMaxFixture)
-    b = SOAPReferences(
-        ["c", "d"], [[2, 2], [3, 3]], nMaxFixture + nshift, lMaxFixture + lshift
-    )
-    if lshift == 0 and nshift == 0:
-        # this must not throw!
-        try:
-            SOAPify.mergeReferences(a, b)
-        except ValueError:
-            pytest.fail("nmax and lmax are not changed: should not throw")
-        return
-    with pytest.raises(ValueError):
-        SOAPify.mergeReferences(a, b)
 
 
 @pytest.fixture(
@@ -271,52 +238,6 @@ def test_fillSOAPVectorFromdscribeArrayOfVectorMultiSpecies(nMaxFixture, lMaxFix
                         assert c[l, n, np] == a[frame, i, limited]
                         assert c[l, np, n] == a[frame, i, limited]
                         limited += 1
-
-
-def test_transitionMatrix(input_mockedTrajectoryClassification):
-    data: SOAPclassification = input_mockedTrajectoryClassification
-    expectedTmat = numpy.zeros((len(data.legend), len(data.legend)))
-    stride = 1
-    for atomID in range(data.references.shape[1]):
-        for frame in range(stride, data.references.shape[0]):
-            expectedTmat[
-                data.references[frame - stride, atomID],
-                data.references[frame, atomID],
-            ] += 1
-
-    tmat = SOAPify.transitionMatrixFromSOAPClassification(data, stride=stride)
-    assert tmat.shape[0] == len(data.legend)
-    assert_array_equal(tmat, expectedTmat)
-
-    tmatNorm = SOAPify.transitionMatrixFromSOAPClassificationNormalized(
-        data, stride=stride
-    )
-    assert tmatNorm.shape[0] == len(data.legend)
-    assert_array_equal(tmatNorm, SOAPify.normalizeMatrix(expectedTmat))
-
-
-def test_residenceTime(input_mockedTrajectoryClassification):
-    data: SOAPclassification = input_mockedTrajectoryClassification
-    expectedResidenceTimes = [[] for i in range(len(data.legend))]
-    for atomID in range(data.references.shape[1]):
-        prevState = data.references[0, atomID]
-        time = 0
-        for frame in range(1, data.references.shape[0]):
-            state = data.references[frame, atomID]
-            if state != prevState:
-                expectedResidenceTimes[prevState].append(time)
-                time = 0
-                prevState = state
-            time += 1
-        # the last state does not have an out transition, appendig negative time to make it clear
-        expectedResidenceTimes[prevState].append(-time)
-
-    for i in range(len(expectedResidenceTimes)):
-        expectedResidenceTimes[i] = numpy.sort(numpy.array(expectedResidenceTimes[i]))
-
-    ResidenceTimes = SOAPify.calculateResidenceTimes(data)
-    for stateID in range(len(expectedResidenceTimes)):
-        assert_array_equal(ResidenceTimes[stateID], expectedResidenceTimes[stateID])
 
 
 def test_centerMaskCreator():
